requirements.txt
torch
torchaudio
lightning

pydantic
pretty_midi
tqdm
mir_eval
pillow